+++ src/flower/tracker/github/iteration.clj
@@ -38,7 +38,7 @@
 
 (defrecord GithubTrackerIteration [tracker it-id it-name it-url it-path it-calendar it-current]
   proto/TrackerIterationProto
-  (get-id [iteration] it-id)
+  (get-iteration-id [iteration] it-id)
   (get-calendar [iteration] it-calendar)
   (get-capacity [iteration] (private-get-github-capacity tracker iteration)))