+++ src/flower/tracker/github/task.clj
@@ -1,7 +1,16 @@
 (ns flower.tracker.github.task
-  (:require [flower.macros :as macros]
+  (:require [clojure.data :as data]
+            [clojure.set :as set]
+            [flower.macros :as macros]
             [flower.tracker.proto :as proto]
             [flower.tracker.github.common :as common]))
+
+;;
+;; Private declarations
+;;
+
+(declare private-set-github-workitem!)
+
 
 ;;
 ;; Public definitions
@@ -10,8 +19,10 @@
 (defrecord GithubTrackerTask [tracker task-id task-title task-type task-state task-tags]
   proto/TrackerTaskProto
   (get-tracker [tracker-task] tracker)
+  (get-task-id [tracker-task] task-id)
   (get-state [tracker-task] task-state)
-  (get-type [tracker-task] task-type))
+  (get-type [tracker-task] task-type)
+  (update! [tracker-task] (private-set-github-workitem! tracker-task)))
 
 
 (macros/public-definition get-github-workitems cached)
@@ -45,3 +56,14 @@
                [:context :tasks-map-function]
                (fn [task] task))
        (private-get-github-workitems-before-map tracker task-ids)))
+
+
+(defn- private-set-github-workitem! [tracker-task]
+  (let [tracker (proto/get-tracker tracker-task)
+        task-id (proto/get-task-id tracker-task)
+        old-workitem (first (proto/get-tasks tracker [task-id]))
+        fields (second (data/diff old-workitem tracker-task))]
+    (common/set-github-workitem-inner! tracker task-id fields)
+    (common/get-github-workitems-inner-clear-cache!)
+    (get-github-workitems-clear-cache!)
+    tracker-task))