+++ src/flower/tracker/gitlab/task.clj
@@ -1,7 +1,16 @@
 (ns flower.tracker.gitlab.task
-  (:require [flower.macros :as macros]
+  (:require [clojure.data :as data]
+            [clojure.set :as set]
+            [flower.macros :as macros]
             [flower.tracker.proto :as proto]
             [flower.tracker.gitlab.common :as common]))
+
+;;
+;; Private declarations
+;;
+
+(declare private-set-gitlab-workitem!)
+
 
 ;;
 ;; Public definitions
@@ -10,8 +19,10 @@
 (defrecord GitlabTrackerTask [tracker task-id task-title task-type task-state task-tags]
   proto/TrackerTaskProto
   (get-tracker [tracker-task] tracker)
+  (get-task-id [tracker-task] task-id)
   (get-state [tracker-task] task-state)
-  (get-type [tracker-task] task-type))
+  (get-type [tracker-task] task-type)
+  (update! [tracker-task] (private-set-gitlab-workitem! tracker-task)))
 
 
 (macros/public-definition get-gitlab-workitems cached)
@@ -32,7 +43,7 @@
                              (.getUsername assignee)
                              nil))
           :task-state (.getState %)
-          :task-tags (.getLabels %)})
+          :task-tags (seq (.getLabels %))})
        (if (empty? task-ids)
          (common/get-gitlab-workitems-inner tracker)
          (common/get-gitlab-workitems-inner tracker task-ids))))
@@ -43,3 +54,14 @@
                [:context :tasks-map-function]
                (fn [task] task))
        (private-get-gitlab-workitems-before-map tracker task-ids)))
+
+
+(defn- private-set-gitlab-workitem! [tracker-task]
+  (let [tracker (proto/get-tracker tracker-task)
+        task-id (proto/get-task-id tracker-task)
+        old-workitem (first (proto/get-tasks tracker [task-id]))
+        fields (second (data/diff old-workitem tracker-task))]
+    (common/set-gitlab-workitem-inner! tracker task-id fields)
+    (common/get-gitlab-workitems-inner-clear-cache!)
+    (get-gitlab-workitems-clear-cache!)
+    tracker-task))