+++ src/flower/tracker/jira/common.clj
@@ -5,7 +5,10 @@
            (com.atlassian.util.concurrent Promises)
            (com.atlassian.jira.rest.client.internal.async AsynchronousHttpClientFactory
                                                           AsynchronousJiraRestClient)
-           (com.atlassian.jira.rest.client.auth BasicHttpAuthenticationHandler)))
+           (com.atlassian.jira.rest.client.auth BasicHttpAuthenticationHandler)
+           (com.atlassian.jira.rest.client.api.domain IssueFieldId)
+           (com.atlassian.jira.rest.client.api.domain.input IssueInputBuilder
+                                                            TransitionInput)))
 
 
 ;;
@@ -18,6 +21,7 @@
 (macros/public-definition get-jira-workitems-inner cached)
 (macros/public-definition get-jira-iterations-inner cached)
 (macros/public-definition get-jira-capacity-inner cached)
+(macros/public-definition set-jira-workitem-inner!)
 
 
 ;;
@@ -69,3 +73,36 @@
 
 (defn- private-get-jira-capacity-inner [tracker iteration]
   nil)
+
+
+(defn- private-set-jira-workitem-inner! [tracker task-id fields]
+  (let [conn-inner (get-jira-conn-inner tracker)
+        issue-client (.getIssueClient conn-inner)
+        workitem-inner (first (get-jira-workitems-inner tracker [task-id]))
+        workitem-input-builder (IssueInputBuilder.)]
+    (if (contains? fields :task-state)
+      (let [transitions (-> issue-client
+                            (.getTransitions workitem-inner)
+                            (.claim))
+            transition (first (filter #(= (get fields :task-state)
+                                          (.getName %))
+                                      transitions))]
+        (if transition
+          (-> issue-client
+              (.transition workitem-inner
+                           (TransitionInput. (.getId transition) nil))
+              (.claim)))))
+    (if (contains? fields :task-title)
+      (.setSummary workitem-input-builder (get fields :task-title)))
+    (if (contains? fields :task-assignee)
+      (.setAssigneeName workitem-input-builder (get fields :task-assignee)))
+    (if (contains? fields :task-tags)
+      (.setFieldValue workitem-input-builder
+                      (.id IssueFieldId/LABELS_FIELD)
+                      (get fields :task-tags)))
+    (if (some true? (map (partial contains? fields)
+                         [:task-title :task-assignee :task-tags]))
+      (-> issue-client
+          (.updateIssue task-id
+                        (.build workitem-input-builder))
+          (.claim)))))